+++ lib/generators/happy_seed/googleoauth/templates/lib/google_api_client.rb
@@ -1,0 +1,7 @@
+class GoogleApiClient
+  def self.client( identity )
+    client = Google::APIClient.new(:application_name => 'HappySeed App', :application_version => "1.0.0" )
+    client.authorization.update_token!({:access_token => identity.accesstoken, :refresh_token => identity.refreshtoken})
+    client
+  end
+end