+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Version 0.0.12
   - Updated rails dependancy to 4.2
   - Moved to guard from autotest
   - Added cucumber
+  - Moved to bh for bootstrap helpers
+  
 
 Version 0.0.10
   - Started the Changelog