+++ CHANGELOG.md
@@ -2,8 +2,11 @@ Version 0.0.12
   - Updated rails dependancy to 4.2
   - Moved to guard from autotest
   - Added cucumber
+  - Added spring to rspec test
   - Moved to bh for bootstrap helpers
-  
+  - Refactored omniauth generators
+  - Updated ActiveAdmin
+  - Added config defaults from ~/.seed_defaults
 
 Version 0.0.10
   - Started the Changelog